feat: Set up initial DSL structure for TextComponentScope and core-api integration
