+++ modules/core/src/main/kotlin/org/eventhorizonlab/kyoriadventuredsl/core/TextComponentScope.kt
@@ -1,16 +1,12 @@
-package org.eventhorizonlab.kyoriadventuredsl
+package org.eventhorizonlab.kyoriadventuredsl.core
 
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.TextComponent
 import net.kyori.adventure.text.format.NamedTextColor
 import net.kyori.adventure.text.format.TextDecoration
-import org.eventhorizonlab.kyoriadventuredsl.api.TextComponentScope
+import org.eventhorizonlab.core.api.TextComponentScope
 
-@DslMarker
-internal annotation class AdventureDsl
-
-@AdventureDsl
-internal class DefaultTextComponentScope(
+internal class TextComponentScope(
     private val builder: TextComponent.Builder
 ) : TextComponentScope {
     override fun content(content: String) {
@@ -27,14 +23,16 @@ internal class DefaultTextComponentScope(
 
     override fun text(init: TextComponentScope.() -> Unit) {
         val childBuilder = Component.text()
-        DefaultTextComponentScope(childBuilder).init()
+        TextComponentScope(childBuilder).init()
         builder.append(childBuilder.build())
     }
 
     fun build() = builder.build()
-}
 
-fun textComponent(init: TextComponentScope.() -> Unit): TextComponent {
-    val builder = Component.text()
-    return DefaultTextComponentScope(builder).apply(init).build()
+    class Factory : TextComponentScope.Factory {
+        override fun create(init: TextComponentScope.() -> Unit): TextComponent {
+            val builder = Component.text()
+            return TextComponentScope(builder).apply(init).build()
+        }
+    }
 }
