+++ modules/core/src/main/resources/META-INF/services/org.eventhorizonlab.core.api.TextComponentScope$Factory
@@ -1,0 +1,1 @@
+org.eventhorizonlab.kyoriadventuredsl.core.TextComponentScope$Factory
